remove cache and fix compatibility with VS code web
