+++ .github/workflows/ci.yml
@@ -20,7 +20,7 @@ jobs:
       uses: actions/setup-node@v4
       with:
         node-version: '20'
-    - run: npm ci
+    - run: npm ci --foreground-scripts
     - run: npm run lint
       if: ${{ matrix.os == 'ubuntu-latest' }}
     - run: npm run build-web
