+++ package.json
@@ -637,7 +637,7 @@
 		"@fontsource/open-sans": "4.5.14",
 		"@fontsource/roboto-mono": "4.5.10",
 		"@highlightjs/cdn-assets": "~11.9.0",
-		"@playwright/test": "~1.44.0",
+		"@playwright/test": "^1.47.2",
 		"@types/lodash.throttle": "~4.1",
 		"@types/mocha": "~10.0",
 		"@types/node": "~20.14",
@@ -677,6 +677,7 @@
 		"@asciidoctor/core": "2.2.7",
 		"@asciidoctor/docbook-converter": "2.0.0",
 		"@orcid/bibtex-parse-js": "0.0.25",
+		"@vscode/ripgrep": "1.15.9",
 		"asciidoctor-kroki": "0.18.1",
 		"html-entities": "^2.4.0",
 		"js-yaml": "^4.1.0",
@@ -686,7 +687,6 @@
 		"util": "^0.12.5",
 		"uuid": "8.3.2",
 		"vscode-nls": "5.2.0",
-		"vscode-ripgrep": "^1.13.2",
 		"vscode-uri": "^3.0.3",
 		"worker-thread": "^1.1.0"
 	},
