+++ src/asciidocEngine.ts
@@ -5,7 +5,8 @@ import { ExtensionContentSecurityPolicyArbiter } from './security'
 import { AsciidocPreviewConfigurationManager } from './features/previewConfig'
 import { SkinnyTextDocument } from './util/document'
 import { AsciidocContributionProvider } from './asciidocExtensions'
-import { AntoraSupportManager, getAntoraDocumentContext, getAntoraConfig } from './features/antora/antoraSupport'
+import { getAntoraDocumentContext, getAntoraConfig } from './features/antora/antoraDocument'
+import { AntoraSupportManager } from './features/antora/antoraContext'
 import { WebviewResourceProvider } from './util/resources'
 import { AsciidoctorConfigProvider } from './features/asciidoctorConfig'
 import { AsciidocTextDocument } from './asciidocTextDocument'
