+++ src/features/asciidoctorExtensions.ts
@@ -2,7 +2,7 @@ import vscode from 'vscode'
 import { AsciidoctorExtensionsSecurityPolicyArbiter } from '../security'
 import { Asciidoctor } from '@asciidoctor/core'
 import { mermaidJSProcessor } from './mermaid'
-import { wrappedFindFiles } from '../util/wrappedFindFiles'
+import { findFiles } from '../util/findFiles'
 
 export interface AsciidoctorExtensionsProvider {
   activate(registry: Asciidoctor.Extensions.Registry): Promise<void>;
@@ -39,7 +39,7 @@ export class AsciidoctorExtensions {
   }
 
   private async getExtensionFilesInWorkspace (): Promise<vscode.Uri[]> {
-    return wrappedFindFiles('.asciidoctor/lib/**/*.js')
+    return findFiles('.asciidoctor/lib/**/*.js')
   }
 
   private isAsciidoctorExtensionsRegistrationEnabled (): boolean {
