+++ src/test/antoraSupport.test.ts
@@ -2,7 +2,7 @@ import os from 'os'
 import * as vscode from 'vscode'
 import * as assert from 'assert'
 import 'mocha'
-import { findAntoraConfigFile, getAntoraDocumentContext } from '../features/antora/antoraSupport'
+import { findAntoraConfigFile, getAntoraDocumentContext } from '../features/antora/antoraDocument'
 import { resetAntoraSupport, createDirectories, createDirectory, createFile, createLink, enableAntoraSupport, removeFiles } from './workspaceHelper'
 import { extensionContext } from './helper'
 import { getDefaultWorkspaceFolderUri } from '../util/workspace'
